backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
mongoengine>=0.29
django-rest-framework-mongoengine>=3.4
PyJWT>=2.9
python-dotenv>=1.0
